Move round advancement buttons to header
